initial
first implementation that will actually be communicating with nimbus.  The current version is non
functional, but very close to working, most of the missing functionality for this to be released
lies with the producer.

Still need to clarify what an art file has, and how an art file links images to itself.